Created some information outlines about what should be impletmented
Created some descriptions about how things should be implemented.

Things that are going to be nessecary to complete are the following:

chess board including chess pieces
theres going to be a general chess piece abstract class,
and then theres going to be the subclasses like the queen etc. they are going to have to interact.

then we are going to need server and client side communication through codes. things like mv [piece id] [location] is going to be nessecary,

we can test that simply through testing and whatnot.

those things are easy though.